+++ .env.template
@@ -5,7 +5,6 @@ DEBUG=True
 SECRET_KEY=
 ALLOWED_HOSTS=0.0.0.0,127.0.0.1,localhost
 DATABASE_URL=postgres://postgres@db:5432/postgres
-REDIS_URL=redis://redis:6379/0
 FYLE_AUTHORISE_URI =
 FYLE_CALLBACK_URI = "http://localhost:8000/workspace/connect/fyle/"
 FYLE_TOKEN_URI =
