+++ .pylintrc
@@ -141,7 +141,7 @@ disable=missing-docstring,
         deprecated-sys-function,
         exception-escape,
         comprehension-escape,
-        broad-except
+        broad-except,
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
