+++ apps/fyle_expense/job_urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 
-from apps.fyle_expense.views import ExpenseGroupTaskView, InvoiceTaskView
+from apps.fyle_expense.views import ExpenseGroupTriggerView, InvoiceTriggerView
 
 urlpatterns = [
-    path('trigger/', ExpenseGroupTaskView.as_view(), name='expense_group_trigger'),
-    path('<group_id>/invoice/trigger/', InvoiceTaskView.as_view(), name='invoice_trigger')
+    path('trigger/', ExpenseGroupTriggerView.as_view(), name='expense_group_trigger'),
+    path('<group_id>/invoice/trigger/', InvoiceTriggerView.as_view(), name='invoice_trigger')
 ]
