+++ apps/fyle_expense/views.py
@@ -13,8 +13,9 @@ from rest_framework.views import APIView
 
 from apps.fyle_expense.models import ExpenseGroup, Expense
 from apps.task_log.models import TaskLog
-from apps.task_log.tasks_test import fetch_expenses_and_create_groups, async_create_invoice_and_post_to_xero, \
+from apps.task_log.tasks import create_invoice_and_post_to_xero, \
     schedule_invoice_creation
+from apps.task_log.tasks import fetch_expenses_and_create_groups
 from apps.xero_workspace.models import CategoryMapping
 
 
@@ -73,20 +74,6 @@ class ExpenseGroupView(View):
             schedule_invoice_creation(workspace_id, expense_group_ids, request.user)
             messages.success(request, 'Resync started successfully. Expenses will be exported soon!')
         return HttpResponseRedirect(self.request.path_info)
-
-
-class ExpenseGroupTaskView(APIView):
-    """
-    Expense Group Task view
-    """
-    http_method_names = ['post']
-
-    def post(self, request, workspace_id):
-        task_log = TaskLog.objects.get(id=request.data.get('task_log_id'))
-
-        fetch_expenses_and_create_groups(workspace_id, task_log, request.user)
-
-        return Response(status=status.HTTP_200_OK)
 
 
 class ExpenseView(View):
@@ -170,9 +157,23 @@ class InvoiceDetailsView(View):
         return JsonResponse(invoice_fields)
 
 
-class InvoiceTaskView(APIView):
+class ExpenseGroupTriggerView(APIView):
     """
-    Invoice task view
+    Expense Group creation job trigger view
+    """
+    http_method_names = ['post']
+
+    def post(self, request, workspace_id):
+        task_log = TaskLog.objects.get(id=request.data.get('task_log_id'))
+
+        fetch_expenses_and_create_groups(workspace_id, task_log, request.user)
+
+        return Response(status=status.HTTP_200_OK)
+
+
+class InvoiceTriggerView(APIView):
+    """
+    Invoice creation job trigger view
     """
     http_method_names = ['post']
 
@@ -182,6 +183,6 @@ class InvoiceTaskView(APIView):
         expense_group = ExpenseGroup.objects.get(id=group_id)
         task_log = TaskLog.objects.get(id=task_log_id)
 
-        async_create_invoice_and_post_to_xero(expense_group, task_log)
+        create_invoice_and_post_to_xero(expense_group, task_log)
 
         return Response(status=status.HTTP_200_OK)
