+++ apps/xero_workspace/tasks.py
@@ -1,5 +1,6 @@
+from apps.task_log.tasks import fetch_expenses_and_create_groups
+
 from apps.task_log.models import TaskLog
-from apps.task_log.tasks_test import fetch_expenses_and_create_groups
 from apps.xero_workspace.utils import connect_to_fyle
 from fyle_jobs import FyleJobsSDK
 from fyle_xero_integration_web_app import settings
