+++ docker-compose.yml
@@ -7,36 +7,19 @@ services:
     container_name: fyle_xero_web_app
     restart: always
     volumes:
-    - ./:/code
+      - ./:/code
     ports:
-    - 8000:8000
+      - 8000:8000
     depends_on:
-    - db
-
-  qcluster:
-    build: .
-    entrypoint: python manage.py qcluster
-    container_name: fyle_xero_qcluster
-    restart: always
-    volumes:
-    - ./:/code
-    depends_on:
-    - db
-    - redis
+      - db
 
   db:
     image: "postgres:12.1-alpine"
     container_name: fyle_xero_db
     ports:
-    - "5432:5432"
+      - "5432:5432"
     volumes:
-    - postgres_data:/var/lib/postgresql/data/
-
-  redis:
-    image: "redis:alpine"
-    container_name: fyle_xero_redis
-    command: [sh, -c, "rm -f /data/dump.rdb && redis-server --save ''"]
-
+      - postgres_data:/var/lib/postgresql/data/
 
 volumes:
   postgres_data:
