+++ requirements.txt
@@ -8,7 +8,6 @@ defusedxml==0.6.0
 Django==2.2.8
 django-allauth==0.41.0
 django-picklefield==2.0
-django-q==1.2.1
 idna==2.8
 isort==4.3.21
 lazy-object-proxy==1.4.1
@@ -22,7 +21,6 @@ python-dateutil==2.8.0
 python-decouple==3.1
 python3-openid==3.1.0
 pytz==2019.2
-redis==3.3.8
 requests==2.22.0
 requests-oauthlib==1.2.0
 six==1.12.0
